refactor(ControlledStepFormDialog): ボタン関連の型をtype.tsに切り出し、ReturnTypeへの依存を解消
StepFormDialogContentInnerがReturnType<typeof useStepFormDialogButton>で型を参照していたため、
useStepFormDialogButtonの実装に依存する形になっていた。type.tsにActualButtonTypeとして
明示的に定義し、そちらを参照するように変更した。
